+++ common/config/bootstrap.php
@@ -1,4 +1,11 @@
 <?php
+// Require core files
+require(__DIR__ . '/../helpers.php');
+
+// Init application constants
+defined('YII_DEBUG') or define('YII_DEBUG', env('YII_DEBUG'));
+defined('YII_ENV') or define('YII_ENV', env('YII_ENV', 'prod'));
+
 // Path aliases
 Yii::setAlias('@base', realpath(__DIR__.'/../../'));
 Yii::setAlias('@common', realpath(__DIR__.'/../../common'));
@@ -9,10 +16,9 @@ Yii::setAlias('@storage', realpath(__DIR__.'/../../storage'));
 Yii::setAlias('@tests', realpath(__DIR__.'/../../tests'));
 
 // Url Aliases
-Yii::setAlias('@frontendUrl', getenv('FRONTEND_URL'));
-Yii::setAlias('@backendUrl', getenv('BACKEND_URL'));
-Yii::setAlias('@storageUrl', getenv('STORAGE_URL'));
+Yii::setAlias('@frontendUrl', env('FRONTEND_URL'));
+Yii::setAlias('@backendUrl', env('BACKEND_URL'));
+Yii::setAlias('@storageUrl', env('STORAGE_URL'));
 
 
-// Other
-require(__DIR__ . '/../shortcuts.php');
+