+++ common/helpers.php
@@ -48,3 +48,29 @@ function activeTextinput($form, $model, $attribute, $inputOptions = [], $fieldOp
 {
     return $form->field($model, $attribute, $fieldOptions)->textInput($inputOptions);
 }
+
+/**
+ * @param string $key
+ * @param mixed $default
+ * @return mixed
+ */
+function env($key, $default = false) {
+
+    $value = getenv($key);
+
+    if ($value === false) {
+        return $default;
+    }
+
+    switch (strtolower($value)) {
+        case 'true':
+        case '(true)':
+            return true;
+
+        case 'false':
+        case '(false)':
+            return false;
+    }
+
+    return $value;
+}